fix converter

+++ DeviantartApi/Objects/Deviation.cs
@@ -145,17 +145,17 @@ namespace DeviantartApi.Objects
             public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
             {
                 var dateTime = (DateTime) value;
-                writer.WriteValue(dateTime.Subtract(new DateTime(1970, 1, 1)).TotalSeconds);
+                writer.WriteValue((long)dateTime.Subtract(new DateTime(1970, 1, 1)).TotalSeconds);
             }
 
             public override bool CanConvert(Type objectType)
             {
-                return objectType == typeof(int);
+                return objectType == typeof(long);
             }
 
             public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
             {
-                var unixTime = (int) reader.Value;
+                var unixTime = (long) reader.Value;
                 var dateTime =
                     new DateTime(1970, 1, 1, 0, 0, 0, 0, System.DateTimeKind.Utc).AddSeconds(unixTime).ToLocalTime();
                 return dateTime;
